requirements.txt
fastapi==0.110.0
uvicorn==0.15.0
requests==2.22.0
beautifulsoup4==4.7.1
lxml==4.7.1
pydantic==2.6.3
aiohttp==3.8.1
gql==3.0.0
google-cloud-pubsub==2.21.1
fastapi-cache2==0.2.1
redis==4.6.0
requests_toolbelt==1.0.0
firebase-admin==6.5.0
pyjwt==2.8.0
python-multipart==0.0.9
pymongo==4.8.0
google-cloud-logging==2.7.0
starlette
pytz
meilisearch